requirements.txt
apipkg==1.4
appdirs==1.4.2
attr==0.2.0
attrs==16.3.0
eradicate==0.1.6
execnet==1.4.1
flake8==3.3.0
mccabe==0.6.1
packaging==16.8
pluggy==0.4.0
py==1.4.32
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.1.10
pytest==3.0.6
pytest-cache==1.0
pytest-eradicate==0.0.3
pytest-flake8==0.8.1
pytest-sugar==0.8.0
python-decouple==3.0
requests==2.13.0
six==1.10.0
termcolor==1.1.0
tox==2.6.0
virtualenv==15.1.0
xmltodict==0.10.2
future==0.16.0
